moved functionality to a new helper
